Proteccion contra SQL Injection y parte de editar perfil que no estaba hecho

+++ src/Model/Lista/InterfazSASLista.java
@@ -14,7 +14,7 @@ public interface InterfazSASLista {
 	public void eliminar(Lista lista, Usuario usuario) throws ErrorAutenticacion, ErrorEliminacion;
 	public void modificar(Lista lista, Usuario usuario) throws ErrorAutenticacion, ErrorGuardado;
 	public void crearLista(Lista lista, Usuario usuario) throws ErrorAutenticacion, ErrorCreacionObjeto, ErrorGuardado;
-	public void crearListaAuto(ListaAuto lista, Usuario usuario, int duracion) throws ErrorAutenticacion, ErrorCreacionObjeto, ErrorConsulta, ErrorGuardado;
+	public void crearListaAuto(ListaAuto lista, Usuario usuario, double duracion) throws ErrorAutenticacion, ErrorCreacionObjeto, ErrorConsulta, ErrorGuardado;
 	public void anadirCancion(Cancion cancion, Lista lista, Usuario usuario) throws ErrorAutenticacion, ErrorCreacionObjeto, ErrorConsulta, ErrorGuardado;
 	public void eliminarCancion(Cancion cancion, Lista lista, Usuario usuario) throws ErrorAutenticacion, ErrorCreacionObjeto, ErrorConsulta, ErrorGuardado;
 	public ArrayList<Lista> mostrar(Usuario usuarioActual) throws ErrorAutenticacion, ErrorConsulta, ErrorCreacionObjeto;
